feat: add sum_impulse_components
Working towards removing the repeated code block in `rtrans`. This
new function, depending on the arguments, should be able to do both the
Newtonian and relativistic cases.

The arguments bit is key -- ideally, all the common arguments would be
packed up into a derived type and handed over as configuration, and only
those parameters that need changing would be changed. But that's a few
refactors away...

Note that the tolerances of some of the tests had to be adjusted here.
There was never more than one value that disagreed, which suggests
perhaps an off-by-one index problems somewhere.